Cleans code up.

+++ org.jhotdraw8.draw/src/main/java/org.jhotdraw8.draw/org/jhotdraw8/graph/SimpleMutableDirectedGraph.java
@@ -228,7 +228,7 @@ public class SimpleMutableDirectedGraph<V, A> extends AbstractDirectedGraphBuild
         if (vidxBox == null) {
             return;
         }
-        int vidx = (int) vidxBox;
+        int vidx = vidxBox;
         // Remove all outgoing vertices
         for (int i = getNextCount(vidx) - 1; i >= 0; i--) {
             int indexOfRemovedArrow = buildRemoveArrowAt(vidx, i);
